+++ Tests/AdvancedMappingCrud.Cosmos.Tests/AdvancedMappingCrud.Cosmos.Tests.Application/Customers/FindCustomerByName/FindCustomerByNameQueryValidator.cs
@@ -1,0 +1,23 @@
+using FluentValidation;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Application.MediatR.FluentValidation.QueryValidator", Version = "2.0")]
+
+namespace AdvancedMappingCrud.Cosmos.Tests.Application.Customers.FindCustomerByName
+{
+    public class FindCustomerByNameQueryValidator : AbstractValidator<FindCustomerByNameQuery>
+    {
+        [IntentManaged(Mode.Merge)]
+        public FindCustomerByNameQueryValidator()
+        {
+            ConfigureValidationRules();
+        }
+
+        private void ConfigureValidationRules()
+        {
+            RuleFor(v => v.Name)
+                .NotNull();
+        }
+    }
+}