+++ Tests/CosmosDB.PrivateSetters/CosmosDB.PrivateSetters.Infrastructure/Repositories/CosmosDBRepositoryBase.cs
@@ -73,6 +73,21 @@ namespace CosmosDB.PrivateSetters.Infrastructure.Repositories
             });
         }
 
+        public async Task<TDomain?> FindAsync(
+            Expression<Func<TDocumentInterface, bool>> filterExpression,
+            CancellationToken cancellationToken = default)
+        {
+            var documents = await _cosmosRepository.GetAsync(AdaptFilterPredicate(filterExpression), cancellationToken);
+
+            if (documents == null || !documents.Any())
+            {
+                return default;
+            }
+            var entity = LoadAndTrackDocument(documents.First());
+
+            return entity;
+        }
+
         public async Task<List<TDomain>> FindAllAsync(CancellationToken cancellationToken = default)
         {
             var documents = await _cosmosRepository.GetAsync(_ => true, cancellationToken);
@@ -99,7 +114,7 @@ namespace CosmosDB.PrivateSetters.Infrastructure.Repositories
             }
         }
 
-        public virtual async Task<List<TDomain>> FindAllAsync(
+        public async Task<List<TDomain>> FindAllAsync(
             Expression<Func<TDocumentInterface, bool>> filterExpression,
             CancellationToken cancellationToken = default)
         {
@@ -109,7 +124,7 @@ namespace CosmosDB.PrivateSetters.Infrastructure.Repositories
             return results;
         }
 
-        public virtual async Task<IPagedList<TDomain>> FindAllAsync(
+        public async Task<IPagedList<TDomain>> FindAllAsync(
             int pageNo,
             int pageSize,
             CancellationToken cancellationToken = default)
@@ -117,7 +132,7 @@ namespace CosmosDB.PrivateSetters.Infrastructure.Repositories
             return await FindAllAsync(_ => true, pageNo, pageSize, cancellationToken);
         }
 
-        public virtual async Task<IPagedList<TDomain>> FindAllAsync(
+        public async Task<IPagedList<TDomain>> FindAllAsync(
             Expression<Func<TDocumentInterface, bool>> filterExpression,
             int pageNo,
             int pageSize,